refactor: move OpenSSH Data helpers to Extensions/Data+Extensions.swift
The two `appendOpenSSHString(_:)` overloads (string and Data variants)
previously lived as a private extension at the bottom of
SecureEnclaveKeyManager.swift with a comment about deferring promotion
until a second call site appears.

Promoting them now to the project's house location for type
extensions: Pulse/Extensions/Data+Extensions.swift, internal access so
the whole app target can use them. The OpenSSH framing is a generic
wire-format primitive, not SSH-key-manager-specific; future protocols
(SCP, the SSH agent protocol, SFTP framing) will want the same helper.

Side benefit: the unit-test commit landing next can exercise the
encoder without re-exporting any private symbols.
