+++ Pulse/SSH/SecureEnclaveKeyManager.swift
@@ -325,24 +325,3 @@ enum SecureEnclaveKeyManager {
     }
 }
 
-// MARK: - OpenSSH framing helper
-//
-// Scoped private to this file: these helpers have a single call site (the OpenSSH
-// wire-format encoder above) and aren't yet general-purpose. Promote to
-// Extensions/Data+Extensions.swift only when a second call site appears outside the
-// SSH subsystem. Premature promotion turns a local helper into an API surface.
-
-private extension Data {
-    /// Appends a length-prefixed string. The length is a big-endian uint32.
-    mutating func appendOpenSSHString(_ string: String) {
-        appendOpenSSHString(Data(string.utf8))
-    }
-
-    /// Appends a length-prefixed binary string. The length is a big-endian uint32.
-    mutating func appendOpenSSHString(_ payload: Data) {
-        var length = UInt32(payload.count).bigEndian
-        // Qualified call: Data's own `withUnsafeBytes(_:)` shadows the free function.
-        Swift.withUnsafeBytes(of: &length) { append(contentsOf: $0) }
-        append(payload)
-    }
-}
